// src/app/models/education.ts
export interface education {
    _id: string
    school_name: string
    city: string
    state: string
    field: string
    percentage: string
    degree_type: string
    graduation_month: string
    graduation_year: Number
}
// school_name: {type: String, required: true},
// city: {type: String, required: true},
// state: {type: String, required: true},
// field: {type: String, required: true},
// percentage: {type: String, required: true},
// degree_type: {type: String, required: true},
// graduation_month: {type: String, required: true},
// graduation_year: {type: Number, required: true},